Add AutoEncoderMixin for autoencoder's history item creation and rating_matrix generation.
