+++ recbole/model/abstract_recommender.py
@@ -109,6 +109,37 @@ class GeneralRecommender(AbstractRecommender):
         self.device = config["device"]
 
 
+class AutoEncoderMixin(object):
+    def build_histroy_items(self, dataset):
+        self.history_item_id, self.history_item_value, _ = dataset.history_item_matrix()
+        self.history_item_id = self.history_item_id
+        self.history_item_value = self.history_item_value
+
+    def get_rating_matrix(self, user):
+        r"""Get a batch of user's feature with the user's id and history interaction matrix.
+
+        Args:
+            user (torch.LongTensor): The input tensor that contains user's id, shape: [batch_size, ]
+
+        Returns:
+            torch.FloatTensor: The user's feature of a batch of user, shape: [batch_size, n_items]
+        """
+        # Following lines construct tensor of shape [B,n_items] using the tensor of shape [B,H]
+        col_indices = self.history_item_id[user].flatten()
+        row_indices = (
+            torch.arange(user.shape[0])
+            .repeat_interleave(self.history_item_id.shape[1], dim=0)
+        )
+        rating_matrix = (
+            torch.zeros(1).repeat(user.shape[0], self.n_items)
+        )
+        rating_matrix.index_put_(
+            (row_indices, col_indices), self.history_item_value[user].flatten()
+        )
+        rating_matrix = rating_matrix.to(self.device)
+        return rating_matrix
+
+
 class SequentialRecommender(AbstractRecommender):
     """
     This is a abstract sequential recommender. All the sequential model should implement This class.
