+++ recbole/model/general_recommender/macridvae.py
@@ -22,13 +22,13 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from recbole.model.abstract_recommender import GeneralRecommender
+from recbole.model.abstract_recommender import AutoEncoderMixin, GeneralRecommender
 from recbole.model.init import xavier_normal_initialization
 from recbole.model.loss import EmbLoss
 from recbole.utils import InputType
 
 
-class MacridVAE(GeneralRecommender):
+class MacridVAE(GeneralRecommender, AutoEncoderMixin):
     r"""MacridVAE is an item-based collaborative filtering model that learns disentangled representations from user
     behavior and simultaneously ranks all items for each user.
 
@@ -51,10 +51,7 @@ class MacridVAE(GeneralRecommender):
         self.std = config["std"]
 
         self.update = 0
-
-        self.history_item_id, self.history_item_value, _ = dataset.history_item_matrix()
-        self.history_item_id = self.history_item_id.to(self.device)
-        self.history_item_value = self.history_item_value.to(self.device)
+        self.build_histroy_items(dataset)
         self.encode_layer_dims = (
             [self.n_items] + self.layers + [self.embedding_size * 2]
         )
@@ -67,30 +64,6 @@ class MacridVAE(GeneralRecommender):
         self.l2_loss = EmbLoss()
         # parameters initialization
         self.apply(xavier_normal_initialization)
-
-    def get_rating_matrix(self, user):
-        r"""Get a batch of user's feature with the user's id and history interaction matrix.
-
-        Args:
-            user (torch.LongTensor): The input tensor that contains user's id, shape: [batch_size, ]
-
-        Returns:
-            torch.FloatTensor: The user's feature of a batch of user, shape: [batch_size, n_items]
-        """
-        # Following lines construct tensor of shape [B,n_items] using the tensor of shape [B,H]
-        col_indices = self.history_item_id[user].flatten()
-        row_indices = (
-            torch.arange(user.shape[0])
-            .to(self.device)
-            .repeat_interleave(self.history_item_id.shape[1], dim=0)
-        )
-        rating_matrix = (
-            torch.zeros(1).to(self.device).repeat(user.shape[0], self.n_items)
-        )
-        rating_matrix.index_put_(
-            (row_indices, col_indices), self.history_item_value[user].flatten()
-        )
-        return rating_matrix
 
     def mlp_layers(self, layer_dims):
         mlp_modules = []
